requirements.txt
Flask==2.1.1
Flask-Admin==1.6.0
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
WTForms==3.0.1
WTForms-Alchemy==0.18.0
arrow==1.2.2
GeoAlchemy2==0.11.1
Shapely==1.8.1.post1
python-dotenv==0.20.0